requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
ultralytics==8.0.196
pillow==10.2.0
httpx==0.27.0
locust==2.31.5
